initial seperation of hooks, colors, size and components
