+++ components/Forms/CallToActionForm.tsx
@@ -1,5 +1,5 @@
 import Box from "@mui/material/Box";
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { useRouter } from "next/router";
 import Input from "@mui/material/Input";
 import Button from "@mui/material/Button";
@@ -7,67 +7,36 @@ import Typography from "@mui/material/Typography";
 import InputLabel from "@mui/material/InputLabel";
 import FormControl from "@mui/material/FormControl";
 import FormHelperText from "@mui/material/FormHelperText";
-
-export const useEmailForm = (): [
-  string,
-  string,
-  (emailValue: string) => void,
-  string,
-  (e: React.MouseEvent) => void,
-  boolean
-] => {
-  const [data, setData] = useState<string>("");
-  const [error, setError] = useState<boolean>(false);
-  const [email, setEmail] = useState<string>("");
-  const [emailLabel, setEmailLabel] = useState<string>("Email Address");
-
-  const handleEmail = (emailValue: string) => {
-    setError(false);
-    setData(emailValue);
-    setEmailLabel(emailLabel);
-  };
-
-  const handleSubmit = (e: React.MouseEvent) => {
-    e.preventDefault();
-    const emailReg = /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/g;
-    if (emailReg.test(data)) {
-      setEmail(data);
-      setData("");
-      return;
-    }
-    setError(true);
-    setEmail("");
-  };
-
-  return [email, data, handleEmail, emailLabel, handleSubmit, error];
-};
+import { useMediaQuery } from "@mui/material";
+import { useEmailForm } from "./CallToActionHook";
 
 const EmailForm: React.FC = () => {
+  const mediumSizeOrHigher = useMediaQuery("(min-width:768px)");
   const router = useRouter();
   const [email, data, handleEmail, emailLabel, handleSubmit, error] =
     useEmailForm();
 
-//   const sendEmailToDb = (email: string) => {
-//     const collectionReference = collection(db, "emails");
-//     addDoc(collectionReference, {
-//       emailAddress: email,
-//       unsubscribe: false,
-//     });
-//     // getDocs(collectionReference) // retrieve all the docs
-//     //   .then((snapshot) => {
-//     //     let emails = [];
-//     //     snapshot.docs.forEach((doc) => {
-//     //       emails.push({ ...doc.data(), id: doc.id });
-//     //     });
-//     //     console.log(emails);
-//     //   })
-//     //   .catch(() => {
-//     //     console.error(" Failed .");
-//     //   });
-//   };
+  //   const sendEmailToDb = (email: string) => {
+  //     const collectionReference = collection(db, "emails");
+  //     addDoc(collectionReference, {
+  //       emailAddress: email,
+  //       unsubscribe: false,
+  //     });
+  //     // getDocs(collectionReference) // retrieve all the docs
+  //     //   .then((snapshot) => {
+  //     //     let emails = [];
+  //     //     snapshot.docs.forEach((doc) => {
+  //     //       emails.push({ ...doc.data(), id: doc.id });
+  //     //     });
+  //     //     console.log(emails);
+  //     //   })
+  //     //   .catch(() => {
+  //     //     console.error(" Failed .");
+  //     //   });
+  //   };
 
-  const sendEmailToDb = (emailAddress: string) :void => {
-    
+  const sendEmailToDb = (emailAddress: string): void => {
+
   }
 
   useEffect(() => {
@@ -84,7 +53,7 @@ const EmailForm: React.FC = () => {
     >
       <Typography variant="h4"> Stay in the loop </Typography>
       <Typography variant="body1">Tune in for roll out date</Typography>
-      <Box component="form" minWidth={'50vh'}>
+      <Box component="form" minWidth={mediumSizeOrHigher ? '100vh' : '50vh'}>
         <FormControl fullWidth error={error} variant="standard">
           <InputLabel htmlFor="component-helper">{emailLabel}</InputLabel>
           <Input
