+++ styles/Home.module.css
@@ -33,6 +33,11 @@
   text-decoration: none;
 }
 
+.subscribeText a {
+  color: #0070f3;
+  text-decoration: none;
+}
+
 .title a:hover,
 .title a:focus,
 .title a:active {
